Flask-Web-App-Tutorial/requirements.txt
flask
Flask-SQLAlchemy
flask-login
flask_wtf
wtforms
werkzeug
ultralytics
psycopg2==2.9.2